+++ components/BottomSheetActions.tsx
@@ -12,7 +12,7 @@ import { useNotificationContext } from "@/contexts";
 import { FirestoreDB } from "@/utils/firestore";
 import { doc, updateDoc } from "firebase/firestore";
 import Toaster from "@/shared-uis/components/toaster/Toaster";
-import { useChatHook } from "@/hooks/use-chat";
+import { useChatHook } from "@/hooks";
 
 interface BottomSheetActionsProps {
   cardType:
