+++ hooks/index.ts
@@ -1,4 +1,6 @@
 import useBreakpoints from "./use-breakpoints";
 import { useStorageState } from "./use-storage-state";
+import useStreamTheme from "./use-stream-theme";
+import { useChatHook } from "./use-chat";
 
-export { useBreakpoints, useStorageState };
+export { useBreakpoints, useStorageState, useChatHook, useStreamTheme };
